pyproject.toml
[tool.poetry]
authors = [
    "Guilherme V. Beira <guilheme.vieira.beira@gmail.com>"
]
name = "web-app"
description = "web-app"
license = "MIT"
version = "0.1.0"

[tool.black]
line-length = 110

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_localfolder = "web_app,apps,tests"
skip_glob = "migrations/*"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
email_validator = "^1.1.0"
prettyconf = "^2.1.0"
uvicorn = "^0.11.5"
gino = {extras = ["starlette"], version = "^1.0.0"}
alembic = "^1.4.2"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-multipart = "^0.0.5"
psycopg2 = "^2.8.6"
aiohttp = "^3.6.3"

[tool.poetry.dev-dependencies]
black = { version = "^19", allow-prereleases = true }
pre-commit = "^1.18"
pytest = "^5.2"
pytest-asyncio = "^0.12"
pytest-cov = "^2.8"
python-status = "i^1.0"
requests = "^2.23.0"
codecov = "^2.0.22"
nest_asyncio = "^1.3.3"
coverage = "^5.1"
pytest-aiohttp = "^0.3.0"
aioresponses = "^0.6.4"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
